move novel list and split actions
