--- src/app/actions/index.js
@@ -1,3 +1,0 @@
-import * as action from './action';
-
-export default action;
