+++ src/app/actions/menuActions.js
@@ -1,0 +1,13 @@
+export const openMenu = () => {
+  return {
+    type: 'OPEN_MENU',
+    payload: true
+  }
+};
+
+export const closeMenu = () => {
+  return {
+    type: 'CLOSE_MENU',
+    payload: true
+  }
+};
