+++ src/app/components/Common/NovelList/NovelCard/NovelCard.js
