+++ src/app/components/Common/NovelList/NovelCard/NovelCardDetail.js
