+++ src/app/components/Common/NovelList/NovelCard/index.js
