+++ src/app/components/Common/NovelList/NovelList.js
