+++ src/app/components/Common/NovelList/index.js
