+++ src/config/routes.js
@@ -1,0 +1,3 @@
+export default Routes = () => {
+  
+}
